+++ src/module/vue/character-sheet.vue
@@ -170,7 +170,6 @@ export default {
   computed: {
     progressItems() {
       return [
-        ...this.actor.items.filter((x) => x.type === 'vow'),
         ...this.actor.items.filter((x) => x.type === 'progress'),
       ]
     },